Crate image_label_tool

image-label-tool

Create bounding boxes around objects in an image and classify them according to your desired labels.

About

This tool is written in Rust and compiled to WebAssembly (Wasm) for use in the browser. It allows you to create bounding boxes around objects in an image and classify them according to your desired labels.

To use the tool, simply open it in your browser and upload an image. You can then draw bounding boxes around the objects you want to classify by clicking and dragging your mouse.

Once you have labeled all the objects in the image, you can export the results. The file will contain the coordinates and labels for each bounding box.

Thank you for using our label tool! I hope it helps with your machine learning projects.

Data Format

Exported data will be stored in Yolo format.

<object-class> <x> <y> <width> <height>

Usage

Clone the repository:

git clone https://github.com/chriamue/image-label-tool
cd image-label-tool

Compile the code to WASM:

wasm-pack build --target web

Run the Web version in your browser

python3 -m http.server

Open your browser on Localhost

Modules

prelude

the image label tool prelude

Functions

init_label_tool

init label tool and start app on given root html element

Type Aliases

Annotation

annotation is an object bounding box in image and class type

Class

object class id