Module image_label_tool::prelude

the image label tool prelude

Re-exports

pub use crate::init_label_tool;

pub use crate::Annotation;

pub use crate::Class;

Structs

AddImageButton

A component representing an “Add Image” button.

AnnotatedImage

Represents an image with associated annotations.

App

The web app

BBox

A bounding box struct with mergre functions

Header

A header component for the Image Label Tool.

ImageStore

A store for managing a list of annotated images.

ImagesList

A component representing a list of images.

LabelTool

struct of label tool that manages annotated images

StatusBar

A status bar component that displays the number of images in the image list panel.

Enums

ImageStoreMsg

Messages for updating the ImageStore.